2010-06-22 09:58 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/src/macro/macro.y   * harbour/src/macro/macro.yyh   * harbour/src/macro/macro.yyc   * harbour/src/macro/macrolex.c     * changed type of valChar.length from int to HB_SIZE

+++ harbour/src/macro/macrolex.c
@@ -228,7 +228,7 @@ static int hb_lexStringExtCopy( YYSTYPE *yylval_ptr, HB_MACRO_PTR pMacro,
          ulLen = pLex->pDst - yylval_ptr->valChar.string;
          *pLex->pDst++ = '\0';
          hb_strRemEscSeq( ( char * ) yylval_ptr->valChar.string, &ulLen );
-         yylval_ptr->valChar.length = ( int ) ulLen;
+         yylval_ptr->valChar.length = ulLen;
          return LITERAL;
       }
       *pLex->pDst++ = ch;
@@ -236,7 +236,7 @@ static int hb_lexStringExtCopy( YYSTYPE *yylval_ptr, HB_MACRO_PTR pMacro,
    ulLen = pLex->pDst - yylval_ptr->valChar.string;
    *pLex->pDst++ = '\0';
    hb_strRemEscSeq( ( char * ) yylval_ptr->valChar.string, &ulLen );
-   yylval_ptr->valChar.length = ( int ) ulLen;
+   yylval_ptr->valChar.length = ulLen;
    hb_macroError( EG_SYNTAX, pMacro );
    return LITERAL;
 }
